+++ src/array/array_builder.rs
@@ -7,6 +7,7 @@ use super::{
     codec::{
         ArrayToArrayCodecTraits, ArrayToBytesCodecTraits, BytesCodec, BytesToBytesCodecTraits,
     },
+    data_type::IncompatibleFillValueError,
     Array, ArrayCreateError, ArrayShape, ChunkGrid, CodecChain, DataType, DimensionName, FillValue,
 };
 
@@ -258,6 +259,13 @@ impl ArrayBuilder {
                     self.shape.len(),
                 ));
             }
+        }
+        if self.data_type.size() != self.fill_value.size() {
+            return Err(IncompatibleFillValueError::new(
+                self.data_type.name(),
+                self.fill_value.clone(),
+            )
+            .into());
         }
 
         Ok(Array {